% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_shellfish.R
\name{clean_shellfish}
\alias{clean_shellfish}
\title{Clean up results of shellfishrisks}
\usage{
clean_shellfish(batch)
}
\arguments{
\item{batch}{}
}
\description{
Clean up results of shellfishrisks
}
\examples{
\dontrun{
clean_shellfish(batch = "dev")
}
}
