% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ctrl_file.R
\name{generate_ctrl_file}
\alias{generate_ctrl_file}
\title{Generate a control file from the default template}
\usage{
generate_ctrl_file(ctrl_file_name = "ctrl", destdir = ".")
}
\arguments{
\item{ctrl_file_name}{the desired name of the control file}

\item{destdir}{the directory to store the control file in}
}
\value{
the file path of the control file
}
\description{
Generate a control file from the default template
}
\examples{
\dontrun{

generate_ctrl_file()
}

}
