% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pypkgs.R
\name{install_pypkgs}
\alias{install_pypkgs}
\title{Install Required Python Packages}
\usage{
install_pypkgs(method = "auto", conda = "auto")
}
\arguments{
\item{method}{python install method}

\item{conda}{path to the conda executable}
}
\description{
Install Required Python Packages
}
\examples{
\dontrun{
install_pylibs()
}

}
