% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shellfish.R
\name{load_shellfish}
\alias{load_shellfish}
\title{load shellfish model}
\usage{
load_shellfish(condaenv = "r-reticulate")
}
\arguments{
\item{condaenv}{the Conda environment used}
}
\value{

}
\description{
load shellfish model
}
\examples{

\dontrun{
load_shellfish()
}


}
