% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shellfish.R
\name{plot_shellfish}
\alias{plot_shellfish}
\title{Plot Shellfish}
\usage{
plot_shellfish(results, type = "rvars")
}
\arguments{
\item{results}{a list of results generated by \code{shellfishrisks::serve_shellfish}}

\item{type}{The type of plot to generate, one of
\itemize{
\item "rvars": plot values from the \code{pop_rvars} objects
\item "fst": plot fst outcomes
\item "popsize": plot the subpopulation sizes
}}
}
\value{
a \code{ggplot2} object
}
\description{
\code{plot_shellfish} returns a \code{ggplot2} object with the desired plot specified by the \code{type} argument.
}
\examples{
\dontrun{

results <- serve_shellfish(batches = c("spam","eggs"))

plot_shellfish(results, type = "rpart")


}

}
