% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_shellfish.R
\name{process_shellfish}
\alias{process_shellfish}
\title{Process Shellfish Genetics Results}
\usage{
process_shellfish(batch, results_dir = ".")
}
\arguments{
\item{batch}{the name of the batch of results}

\item{results_dir}{the name of the directory where results are stored}
}
\value{
a list containing each of the results objects
}
\description{
Process Shellfish Genetics Results
}
\examples{
\dontrun{

results <- process_shellfish(batch = "dev")

}
}
