% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ctrl_folder.R
\name{run_ctrl_folder}
\alias{run_ctrl_folder}
\title{Run all control files in a folder}
\usage{
run_ctrl_folder(ctrl_folder, reps = 1, cores = 1)
}
\arguments{
\item{ctrl_folder}{the location of the folder containing the control files}

\item{reps}{the number of reps per control file}

\item{cores}{the number of cores. Set to greater than 1 to run in parallel}
}
\description{
Run all control files in a folder
}
\examples{
\dontrun{

run_ctrl_folder("my_control_files", cores = 6)

}

}
