% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shellfish_ctrlfile.R
\name{run_shellfishrisk_ctrlfile}
\alias{run_shellfishrisk_ctrlfile}
\title{Run shellfishrisk from control file}
\usage{
run_shellfishrisk_ctrlfile(
  batch = "demo",
  reps = 1,
  coreid = 1,
  ctrl_file_path
)
}
\arguments{
\item{batch}{the name of the batch}

\item{reps}{the number of reps to run}

\item{coreid}{no id}

\item{ctrl_file_path}{the file path to the control file to be used}
}
\value{
nothing
}
\description{
Run shellfishrisk from control file
}
