% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_shellfish.R
\name{serve_shellfish}
\alias{serve_shellfish}
\title{Load and Serve Shellfish Genetics Results}
\usage{
serve_shellfish(batch, results_dir = ".")
}
\arguments{
\item{batch}{the name of the batch of results}

\item{results_dir}{the name of the directory where results are stored}
}
\value{
a list containing each of the results objects
}
\description{
Load and Serve Shellfish Genetics Results
}
\examples{
\dontrun{

results <- serve_shellfish(batch = "dev")

}
}
