% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shellfish.R
\name{theme_shellfish}
\alias{theme_shellfish}
\title{ggplot2 theme for shellfishrisks}
\usage{
theme_shellfish(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
ggplot2 theme for shellfishrisks
}
\examples{

ggplot(mtcars, aes(mpg)) + geom_histogram() + shellfishrisks::theme_shellfish()

}
